requirements.txt
fastapi==0.103.0
uvicorn==0.23.0
sqlalchemy[asyncio]==2.0.20
aiosqlite==0.19.0
openai==1.90.0
python-dotenv==1.0.0
aiofiles==23.2.1
python-multipart==0.0.20
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pytest
httpx
pydantic